Remove unnecessary tower rendering files. Move PixiTowerRenderer.
