--- src/js/client/renderers/towers/CannonRenderer.js
@@ -1,8 +1,0 @@
-
-import { autorun } from 'mobx'
-
-import TowerRenderer from '../TowerRenderer'
-
-export default class CannonRenderer extends TowerRenderer {
-
-}
