--- src/js/client/renderers/towers/FlamethrowerRenderer.js
@@ -1,8 +1,0 @@
-
-import { autorun } from 'mobx'
-
-import TowerRenderer from '../TowerRenderer'
-
-export default class FlamethrowerRenderer extends TowerRenderer {
-
-}
