+++ src/js/client/renderers/towers/PixiCannonRenderer.js
@@ -2,10 +2,8 @@
 import { autorun } from 'mobx'
 
 import { GAME_REFRESH_RATE } from '../../../appConstants'
-import PixiTowerRenderer from '../PixiTowerRenderer'
+import PixiTowerRenderer from './PixiTowerRenderer'
 
 export default class PixiCannonrRenderer extends PixiTowerRenderer {
-
-
 
 }
