+++ src/js/client/renderers/towers/PixiFlamethrowerRenderer.js
@@ -2,7 +2,7 @@
 import { autorun } from 'mobx'
 
 import { GAME_REFRESH_RATE } from '../../../appConstants'
-import PixiTowerRenderer from '../PixiTowerRenderer'
+import PixiTowerRenderer from './PixiTowerRenderer'
 
 export default class PixiFlamethrowerRenderer extends PixiTowerRenderer {
