+++ src/js/client/renderers/towers/index.js
@@ -1,8 +1,6 @@
 
-import CannonRenderer from './CannonRenderer'
 import PixiCannonRenderer from './PixiCannonRenderer'
-import FlamethrowerRenderer from './FlamethrowerRenderer'
 import PixiFlamethrowerRenderer from './PixiFlamethrowerRenderer'
 
 
-export { CannonRenderer, PixiCannonRenderer, FlamethrowerRenderer, PixiFlamethrowerRenderer }
+export { PixiCannonRenderer, PixiFlamethrowerRenderer }
